requirements.txt
pipecat-ai[deepgram,openai,silero,local]==0.0.36
python-dotenv>=1.0.0
openai>=1.27.0
elevenlabs>=1.2.0
